Day3
I'm not super sure about this way of handling input. Solution starts with presumption that we should map Santa movmenet to 2D coordinate system. If Santa moves up we move him 1 point on y-axis, if he moves left we move him -1 on x-axis.

Whenever present is delivered we save it in a dictinary where key is house coordinate and value is number of presents (time visited) delivered. Rest of the problems are just solved by checking how many items there are in the dictionary (because each house has unique coordinate point)