Day5
Solved second problem. Replaced `hasOccuringLetters` with more generic `hasOccuringLettersWithPlaceOffset` because it allows to define with which offset should we search for words. Second part is solved by simple "mechanical" work, we iterate through array and compare characters (with offset), or in case of `containsRepeatingPairs` we walk through string, taking pairs (substring) and searching rest of the string (from pair to the end of original string) for occurrence of pair substring