day 7 - refactoring parsing input
Moved input to separate source file to make things more readable. Switched from parsing input by popping to something more readable (Still think this is not the best way to do it). Next thing to do is to think of a way to construct a wire model (find a way to run the simulation and to know at every step what are values on each wire)