Search algorithm
Implemented recursive search algorithm for a circuit. We run down the (input) circuit and simply look if the output wire has inputs as values or other wires. If we find a value than we apply the (bitwise) operation and return the result. If we find a wire then we try (by the same algorithm) to find a value for that output wire. Basically, we just spin in circles trying to get values as inputs.